feat: full env editor on cloud edit page
